Bug 1756587. Fix a couple of aInnerWindowId parameters to be uint64_t in ImageFactory.cpp. r=aosmond
Depends on D139370

Differential Revision: https://phabricator.services.mozilla.com/D139371

UltraBlame original commit: a714e906b18f5ae80fb22c0ed55a021743bc2ccf

+++ image/ImageFactory.h
@@ -43,7 +43,7 @@ class ImageFactory {
                                              ProgressTracker* aProgressTracker,
                                              const nsCString& aMimeType,
                                              nsIURI* aURI, bool aIsMultiPart,
-                                             uint32_t aInnerWindowId);
+                                             uint64_t aInnerWindowId);
   
 
 
@@ -71,12 +71,12 @@ class ImageFactory {
   static already_AddRefed<Image> CreateRasterImage(
       nsIRequest* aRequest, ProgressTracker* aProgressTracker,
       const nsCString& aMimeType, nsIURI* aURI, uint32_t aImageFlags,
-      uint32_t aInnerWindowId);
+      uint64_t aInnerWindowId);
 
   static already_AddRefed<Image> CreateVectorImage(
       nsIRequest* aRequest, ProgressTracker* aProgressTracker,
       const nsCString& aMimeType, nsIURI* aURI, uint32_t aImageFlags,
-      uint32_t aInnerWindowId);
+      uint64_t aInnerWindowId);
 
   
   virtual ~ImageFactory() = 0;
